+++ blobxfer/models/upload.py
@@ -470,6 +470,16 @@ class Descriptor(object):
         return self.entity.mode == blobxfer.models.azure.StorageModes.Append
 
     @property
+    def remote_is_block_blob(self):
+        # type: (Descriptor) -> bool
+        """Remote destination is an Azure Block Blob
+        :param Descriptor self: this
+        :rtype: bool
+        :return: remote is an Azure Block Blob
+        """
+        return self.entity.mode == blobxfer.models.azure.StorageModes.Block
+
+    @property
     def is_one_shot_block_blob(self):
         # type: (Descriptor) -> bool
         """Is one shot block blob
@@ -477,8 +487,7 @@ class Descriptor(object):
         :rtype: bool
         :return: if upload is a one-shot block blob
         """
-        return (self._ase.mode == blobxfer.models.azure.StorageModes.Block and
-                self._total_chunks == 1)
+        return self.remote_is_block_blob and self._total_chunks == 1
 
     @property
     def requires_put_block_list(self):
@@ -488,8 +497,7 @@ class Descriptor(object):
         :rtype: bool
         :return: if finalize requires a put block list
         """
-        return (self._ase.mode == blobxfer.models.azure.StorageModes.Block and
-                self._total_chunks > 1)
+        return self.remote_is_block_blob and self._total_chunks > 1
 
     @property
     def requires_non_encrypted_md5_put(self):
@@ -512,6 +520,17 @@ class Descriptor(object):
         """
         return (not self.entity.is_encrypted and self.must_compute_md5 and
                 self.remote_is_file)
+
+    @property
+    def requires_access_tier_set(self):
+        # type: (Descriptor) -> bool
+        """Remote destination requires an access tier set operation
+        :param Descriptor self: this
+        :rtype: bool
+        :return: access tier is set
+        """
+        return (self.remote_is_block_blob and
+                self.entity.access_tier is not None)
 
     def requires_resize(self):
         # type: (Descriptor) -> tuple